+++ CHANGELOG.md
@@ -19,6 +19,7 @@ limitations under the License.
 ## 0.7.4 (unreleased)
 - new: decoupled mode configuration in Triton Model Config
 - change: in TensorRT runner, when TensorType.TORCH is the return type lazily convert tensort to Torch
+- change: move from Polygraphy CLI to Polygraphy Python API
 
 
 ## 0.7.3